generatorQR/requirements.txt
# Core Django + REST
django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1

# Base de datos
psycopg2-binary==2.9.9
django-filter==24.2

# Generación de QR
qrcode[pil]==7.4.2
pillow==10.3.0

# Utilidades
requests==2.31.0
python-decouple==3.8
